only build fulltext index for mysql dialect
The DDL instructions to extend the schema only work on MySQL, so when
they are run for other database dialects they fail. This causes tests
that use databases like sqlite to fail, even if they are not using the
full text searching features. This change configures the DDL
instruction to only run when the database dialect is mysql to prevent
these failures.

Signed-off-by: Doug Hellmann <doug@doughellmann.com>